admin: Add an esp-upgrade subcommand
The files in the EFI System Partition (ESP) are never updated by OSTree so
users will always have the same ESP that was created during installation.

This happens because the ESP uses a vfat filesystem, which doesn't support
symbolic links so the update can't be atomic and part of the transaction.

But since the ESP contains the bootloader components (shim, grub, etc) and
other important EFI binaries like the ones needed for firmware updates, it
should be a way to update the ESP with the latest version of the binaries.

These components are stored in the /usr/lib/ostree-boot/efi directory of a
deployment due rpm-ostree replacing the /boot path to /usr/lib/ostree-boot.

Add a ostree-admin-esp-upgrade subcommand, that just copies the content of
/usr/lib/ostree-boot/efi to /boot/efi, so users can update the ESP and get
the latest version of the files that are present in the booted deployment.

Closes: #1649